final refactoring -- Helpers to inner static classes
